requirements.txt
# AWS SDK
boto3>=1.34.0
botocore>=1.34.0

# Azure SDK
azure-identity>=1.15.0
azure-mgmt-compute>=30.5.0
azure-mgmt-containerinstance>=10.1.0
azure-mgmt-containerservice>=28.0.0
azure-mgmt-web>=7.2.0
azure-mgmt-batch>=17.0.0
azure-mgmt-resource>=23.0.0

# Google Cloud SDK
google-cloud-compute>=1.15.0
google-cloud-container>=2.37.0
google-cloud-run>=0.10.0
google-cloud-functions>=1.13.0
google-cloud-appengine-admin>=1.10.0
google-cloud-dataflow-client>=0.8.6
google-auth>=2.27.0

# CLI and output formatting
click>=8.1.7
tabulate>=0.9.0
colorama>=0.4.6

# Utilities
python-dateutil>=2.8.2
requests>=2.31.0

# Development and testing
pytest>=7.4.3
pytest-cov>=4.1.0
pytest-mock>=3.12.0
flake8>=7.0.0
pylint>=3.0.3
black>=23.12.1
